+++ packages/analytics/analytics-chart/src/types/index.ts
@@ -1,5 +1,6 @@
 export * from './chart-data'
 export * from './chart-export'
+export * from './chart-types'
 export * from './chartjs-options'
 export * from './explore-to-dataset-deps'
 export * from './dataset-generation-types'