+++ packages/analytics/dashboard-renderer/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.26.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/dashboard-renderer@0.25.13...@kong-ui-public/dashboard-renderer@0.26.0) (2024-10-03)
+
+
+### Features
+
+* **analytics:** remove `ChartTypes` enum [MA-2708] ([#1688](https://github.com/Kong/public-ui-components/issues/1688)) ([41934de](https://github.com/Kong/public-ui-components/commit/41934deb022449fd173cc7886e4cd299b82a4606))
+
+
+### BREAKING CHANGES
+
+* **analytics:** Remove `ChartTypes` enum
+
+- Unify chart types between `AnalyticsChart` and `DashboardRenderer`.
+- Standardize naming of timeseries line chart.
+- Add union types for available chart types.
+
+
+
+
+
 ## [0.25.13](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/dashboard-renderer@0.25.12...@kong-ui-public/dashboard-renderer@0.25.13) (2024-10-02)
 
 **Note:** Version bump only for package @kong-ui-public/dashboard-renderer